# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rota Certa Passagens - API do site público
  version: 0.1.0
  description: Contrato da API do site público e do painel administrativo.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: rc_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
    cronToken:
      type: http
      scheme: bearer
      description: NOTIFICATIONS_CRON_TOKEN, for unattended scheduler calls with no session cookie.
...
paths:
...
  /api/admin/finance/cost-centers:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Lista centros de custo do módulo financeiro (Fase 1)
      responses:
        "200": { description: Lista de centros de custo }
        "401": { description: Sem sessão válida }
        "403": { description: Papel master necessário }
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Cria um centro de custo
      responses:
        "201": { description: "{ id }" }
        "400": { description: Corpo inválido }
        "409": { description: Nome já usado por outro centro de custo }
  /api/admin/finance/cost-centers/{id}:
    patch:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Atualiza nome ou ativa/desativa um centro de custo (nunca DELETE)
      responses:
        "200": { description: "{ ok: true }" }
        "404": { description: Não encontrado }
        "409": { description: Nome já usado por outro centro de custo }
  /api/admin/finance/categories:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Lista categorias financeiras hierárquicas (receita, custo direto, despesa operacional)
      responses:
        "200": { description: Lista de categorias }
        "401": { description: Sem sessão válida }
        "403": { description: Papel master necessário }
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Cria categoria; categoria-filha deve ter o mesmo "kind" da categoria-pai
      responses:
        "201": { description: "{ id }" }
        "422": { description: Categoria-pai ou centro de custo padrão inexistente, ou kind incompatível com o pai }
  /api/admin/finance/categories/{id}:
    patch:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Atualiza ou desativa uma categoria (bloqueado se houver categoria-filha ativa)
      responses:
        "200": { description: "{ ok: true }" }
        "404": { description: Não encontrado }
        "409": { description: Categoria tem subcategoria ativa; desative-as primeiro }
  /api/admin/finance/accounts:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Lista contas financeiras internas (banco/dinheiro/cartão/carteira) — sem integração bancária real
      responses:
        "200": { description: Lista de contas }
        "401": { description: Sem sessão válida }
        "403": { description: Papel master necessário }
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Cria conta financeira interna
      responses:
        "201": { description: "{ id }" }
        "400": { description: Corpo inválido }
  /api/admin/finance/accounts/{id}:
    patch:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Atualiza ou desativa uma conta financeira (nunca DELETE)
      responses:
        "200": { description: "{ ok: true }" }
        "404": { description: Não encontrado }
  /api/admin/finance/counterparties:
    get:
      tags: [Finance]
      security: [{ sessionCookie: [] }]
      summary: Lista contrapartes (fornecedor, companhia, consolidadora, fornecedor de milhas) — distinta de /api/admin/partners (parceiro de indicação)
      responses:
        "200": { description: Lista de contrapartes }
        "401": { description: Sem sessão válida }
        "403": { description: Papel master necessário }
    post:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Cria contraparte financeira
      responses:
        "201": { description: "{ id }" }
        "400": { description: Corpo inválido }
  /api/admin/finance/counterparties/{id}:
    patch:
      tags: [Finance]
      security: [{ sessionCookie: [], csrfToken: [] }]
      summary: Atualiza ou desativa uma contraparte (nunca DELETE)
      responses:
        "200": { description: "{ ok: true }" }
        "404": { description: Não encontrado }
...